requirements.txt
asgiref==3.4.1
Django==3.2.4
django-prometheus==2.1.0
djangorestframework==3.13.1
prometheus-client==0.12.0
pytz==2021.3
sqlparse==0.4.2
